Simplify declaration of original positions
This way we don't have to declare them every loop, and it is perhaps
a bit dryer this way.

Also make sure that the positions properly resets when you reset the
scene, and make the rotation end a bit earlier so that you can
appreciate the result a bit before it dissappears.